cleaned up a bunch of compiler warnings

+++ source/content/bullet.hpp
@@ -17,7 +17,7 @@ namespace game::content {
             void instr_type_set_relative(int);
             void instr_angle(float);
             //  needs to inform the instruction loop that an element was removed
-            int instr_stop(uint32_t);  
+            size_t instr_stop(uint32_t);  
             void instr_start(uint32_t);
             void instr_frameTrigger(uint32_t, uint32_t);
             void instr_frameTriggerOffset(uint32_t, uint32_t);
